+++ components/onboarder-supplier/back/src/Domain/Mailer/SendEmail.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Akeneo\OnboarderSerenity\Supplier\Domain\Mailer;
+
+use Akeneo\OnboarderSerenity\Supplier\Domain\Mailer\ValueObject\Email;
+
+interface SendEmail
+{
+    public function __invoke(Email $email): void;
+}