+++ components/onboarder-supplier/back/tests/Unit/Infrastructure/Authentication/ContributorAccount/BuildTwigWelcomeEmailTest.php
@@ -1,0 +1,50 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Akeneo\OnboarderSerenity\Supplier\Test\Unit\Infrastructure\Authentication\ContributorAccount;
+
+use Akeneo\OnboarderSerenity\Supplier\Infrastructure\Authentication\ContributorAccount\BuildTwigWelcomeEmail;
+use PHPUnit\Framework\TestCase;
+use Twig\Environment;
+
+final class BuildTwigWelcomeEmailTest extends TestCase
+{
+    /** @test */
+    public function itBuildsAWelcomeEmail(): void
+    {
+        $contributorEmail = 'jeanjacques@example.com';
+        $domain = 'http://wwww.example.com';
+
+        $twig = $this->createMock(Environment::class);
+        $twig
+            ->expects($this->exactly(2))
+            ->method('render')
+            ->withConsecutive(
+                [
+                    '@AkeneoOnboarderSerenitySupplier/Email/contributor-invitation.html.twig',
+                    [
+                        'contributorEmail' => $contributorEmail,
+                        'url' => 'http://wwww.example.com/onboarder/supplier/index.html#/set-up-password/foo',
+                    ],
+                ],
+                [
+                    '@AkeneoOnboarderSerenitySupplier/Email/contributor-invitation.txt.twig',
+                    [
+                        'contributorEmail' => $contributorEmail,
+                        'url' => 'http://wwww.example.com/onboarder/supplier/index.html#/set-up-password/foo',
+                    ],
+                ],
+            )
+            ->willReturnOnConsecutiveCalls(
+                'htmlContent',
+                'textContent',
+            );
+
+        $sut = new BuildTwigWelcomeEmail($twig, $domain);
+        $emailContent = ($sut)('foo', $contributorEmail);
+
+        static::assertSame('htmlContent', $emailContent->htmlContent);
+        static::assertSame('textContent', $emailContent->textContent);
+    }
+}